requirements.txt
pytorch 
tiktoken